test: restore full init coverage metrics
- Add targeted preset edge-case tests in guided init
- Cover missing presets payload and list type guards
- Split interactive prompt IO into lazy-loaded module
- Keep init coverage and runtime checks fast and green

Made-with: Cursor
